Windows_Agent/requirements-clean.txt
appopener==1.7
ddgs==9.10.0
edge_tts==7.0.0
edge_tts==7.2.7
googlesearch_python==1.3.0
groq==1.0.0
keyboard==0.13.5
mtranslate==1.8
Pillow==12.1.0
protobuf==6.33.2
psutil==7.0.0
psutil==7.2.1
pygame==2.6.1
PyQt5==5.15.11
PyQt5_sip==12.17.0
pyqt5_sip==12.17.2
python-dotenv==1.2.1
pywhatkit==5.4
Requests==2.32.5
rich==14.2.0
selenium==4.39.0
webdriver_manager==4.0.2
